requirements.txt
flask
tensorflow==2.10.0
opencv-python
pillow
numpy
